HEEDLS-492 Move All delegates and View delegates viewmodels into folders
