+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/ViewDelegate/WelcomeEmailSentViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates
+﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.ViewDelegate
 {
     using DigitalLearningSolutions.Data.Models.User;
