+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/AllDelegates/AllDelegateItems.cshtml
@@ -1,4 +1,5 @@
 ﻿@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.AllDelegates
 @model AllDelegateItemsViewModel
 
 @{
