+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/AllDelegates/Index.cshtml
@@ -1,6 +1,7 @@
 ﻿@inject IConfiguration Configuration
 @using DigitalLearningSolutions.Web.Models.Enums
 @using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.AllDelegates
 @using Microsoft.Extensions.Configuration
 @model AllDelegatesViewModel
