+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/AllDelegates/_SearchableDelegateCard.cshtml
@@ -1,5 +1,6 @@
 ﻿@using DigitalLearningSolutions.Web.ViewModels.Common
 @using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.AllDelegates
 @model SearchableDelegateViewModel
 
 <div class="searchable-element nhsuk-expander-group" id="@Model.DelegateInfo.Id-card">
