+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/ViewDelegate/WelcomeEmailSent.cshtml
@@ -1,5 +1,6 @@
 ﻿@inject IConfiguration Configuration
 @using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates
+@using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.ViewDelegate
 @using Microsoft.Extensions.Configuration
 @model WelcomeEmailSentViewModel
